data_fetcher/requirements.txt
pandas
yfinance
ta==0.10.2
numpy
pymongo[srv]
google-cloud-secret-manager
functions-framework